backend/requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.5.0
django-cors-middleware==1.5.0
django-heroku==0.3.1
djangorestframework==3.11.1
gunicorn==20.0.4
psutil==5.7.2
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
